Rename CacheableGraphqlValue to CacheableLeafValue, fix comment standards
