+++ src/GraphQL/Field/Entity/EntityIdField.php
@@ -4,7 +4,7 @@ namespace Drupal\graphql\GraphQL\Field\Entity;
 
 use Drupal\Core\Cache\CacheableMetadata;
 use Drupal\Core\Entity\EntityInterface;
-use Drupal\graphql\GraphQL\CacheableGraphqlValue;
+use Drupal\graphql\GraphQL\CacheableLeafValue;
 use Drupal\graphql\GraphQL\Field\FieldBase;
 use Youshido\GraphQL\Execution\ResolveInfo;
 use Youshido\GraphQL\Type\NonNullType;
@@ -36,11 +36,11 @@ class EntityIdField extends FieldBase {
    * @param \Youshido\GraphQL\Execution\ResolveInfo $info
    *   The context information to resolve.
    *
-   * @return CacheableGraphqlValue The associated entity id as a string.
+   * @return CacheableLeafValue The associated entity id as a string.
    *   The associated entity id as a string.
    */
   public function resolve($value, array $args, ResolveInfo $info) {
-    return new CacheableGraphqlValue($value->getEntityType(), [$value]);
+    return new CacheableLeafValue($value->getEntityType(), [$value]);
   }
 
 }
